sfeed_mbox: remove prefix in subject
The prefix in the subject can be too much/redundant if custom labels are used
in the mail client for example.  Filtering or custom labels can be made using
the X-Feedname header.

For mutt it can be useful to use the header X-Label and use the %y format
specified to indicate the label. The index_format is described in the muttrc
man page.

An example:

set index_format = "%4C %Z %[%a, %b %d %H:%M] %?y?%-15.15y ?%-15.15L (%?l?%4l&%4c?) %s"

Other programs have similar headers: Dovecat has "X-Keywords" and Thunderbird
"X-Mozilla-Keys".