fix counting due to uninitialized variable when the time could not be parsed
Since commit 276d5789fd91d1cbe84b7baee736dea28b1e04c0 if the time is empty or
could not be parsed then it is shown/aligned as a blank space instead of being
skipped.

An oversight in this change was that items should be counted and set in
`isnew`.

This commit fixes the uninitialized variable and possible miscounting.