parse categories, add multiple field values support (for categories)
Fields with multiple values are separated by '|'. In the future multiple
enclosure support might be added.

The categories tags are now parsed. This feature is useful for filtering and
categorizing.

Parsing of nested tags such as <author><name> has been improved. This code has
been refactored.

RSS <guid> isPermaLink is now handled differently also and will now prefer a
permalink with "true" (link) over the ID. In practise multiple <guid> in an
item does not happen.