sfeed_web: improve parsing a <link> if it has no type attribute
This happens because the previous link type is not reset when a <link> tag
starts again, but it is reset when a type attribute starts.

Found on the spanish newspaper site: elpais.com

Input:

	<link rel="alternate" href="https://feeds.elpais.com/mrss-s/pages/ep/site/elpais.com/portada" type="application/rss+xml" title="RSS de la portada de El País"/>
	<link rel="canonical" href="https://elpais.com"/>

Would print (second line is incorrect).

	https://feeds.elpais.com/mrss-s/pages/ep/site/elpais.com/portada	application/rss+xml
	https://elpais.com/	application/rss+xml

Now prints:

	https://feeds.elpais.com/mrss-s/pages/ep/site/elpais.com/portada	application/rss+xml

Fix: reset it also at the start of a <link> tag in this case (for <base href />
it is still not wanted).