sfeed_xmlenc: be more paranoid in printing encoding names
sfeed_xmlenc is used automatically in sfeed_update for detecting the encoding.

In particular do not allow slashes anymore either.  For example "//IGNORE" and
"//TRANSLIT" which are normally allowed.

Some iconv implementation might allow other funky names or even pathnames too,
so disallow that.

See also the notes about the "frommap" for the "-f" option.
https://pubs.opengroup.org/onlinepubs/9699919799/utilities/iconv.html

+ some minor parsing handling improvements.