xmlencode: optimize common character output function
Use putc instead of fputc, it can be optimized to macros.

From the OpenBSD man page:

"    putc() acts essentially identically to fputc(), but is a macro that
     expands in-line.  It may evaluate stream more than once, so arguments
     given to putc() should not be expressions with potential side effects."

sfeed_atom, sfeed_frames and sfeed_html are using this function.

Mini-benchmarked sfeed_html and it went from 1.45s to 1.0s with feed files in
total 250k lines (+- 350MB). Tested with clang and gcc on OpenBSD on an older
laptop.