README: fix xargs -P example when there are no feeds
Kindof a non-issue but if theres a sfeedrc with no feeds then xargs will still
be executed and give an error. The xargs -r option (GNU extension) fixes this:

From the OpenBSD xargs(1) man page:
"-r  Do not run the command if there are no arguments.  Normally the
command is executed at least once even if there are no arguments."

Reproducable with the sfeedrc:

	feeds() {
		true
	}