sfeed: small optimization
For feeds with lots of content data:

Small performance improvement (~2%) on systems that implement putchar as a
macro.  On some systems using a function call for putchar it can be easier to
replace with putchar_unlocked.

(On an older MIPS32 VM changing putchar to putchar_unlocked makes writing 5x
faster).