stricter error checking in file streams (input, output)
This also makes the programs exit with a non-zero status when a read or write
error occurs.

This makes checking the exit status more reliable in scripts.

A simple example to simulate a disk with no space left:

	curl -s 'https://codemadness.org/atom.xml' | sfeed > f
	/mnt/test: write failed, file system is full
	echo $?
	0

Which now produces:

	curl -s 'https://codemadness.org/atom.xml' | sfeed > f
	/mnt/test: write failed, file system is full
	write error: <stdout>
	echo $?
	1

Tested with a small mfs on OpenBSD, fstab entry:

	swap /mnt/test mfs rw,nodev,nosuid,-s=1M 0 0