util.h: slightly improve portability
Tested with the scc compiler which is a pure c99 compiler.

sys/types.h is not needed here anymore (it was used for ssize_t).

Side-note: scc can now compile the sfeed parser program!
It requires these changes at the time of writing: Add a strcasecmp and
strncasecmp function and use getchar instead of getchar_unlocked.