sfeed_curses: ignore SIGCHLD only for non-interactive programs
This is a regression from the introduced change.
Else wait(&status) returned -1 and status was uninitialized.

The status of the returned program in the markread() function is used to
visually show it as read/unread. If the program failed it is assumed the
program could not mark it and so it is visually unchanged.

Just to be sure also initialize status to -1 (which can never happen normally)
since the exitstatus range is 0-255.

https://man.openbsd.org/wait#ERRORS
[ECHILD]: "No status from the terminated child process is available because the calling
process has asked the system to discard such status by ignoring the signal
SIGCHLD or setting the flag SA_NOCLDWAIT for that signal."