sfeed_curses: improve waiting on processes and reaping them
- Wait on the exact process id and get its status.

- Handle SIGCHLD explicitly and reap zombie children: ignoring them by using
  sigaction(SIGCHLD, &sa, NULL); would be racy in this case because sfeed_curses
  has interactive and non-interactive programs.

Note while testing: if the markread program would be slow and in the meantime a
plumb process would exit. This signal is now pending and is a zombie process
until the SIGCHLD signal can be processed. This is fine.

This also fixes a regression from commit
30a70fa2dab1925b0eaea04f67e3f86b360386dd because SIGCHLD was ignored in the
parent for interactive processes aswell.  This broke reading the exit status of
the markread program, reproducable by plumbing an item and then trying to mark
it as read (with SFEED_URL_FILE set).