sfeed_curses: close stdin for spawning a plumb program in non-interactive mode
This is only for plumbing in non-interactive mode in forkexec(), but not piping
content.

Probably obvious what the descriptors are, but also add a few comments to dup2
of the file descriptors (stdin, stdout, stderr).

To reproduce a behaviour:

	Plumb script:

	#!/bin/sh
	dmenu

	Then launch it:

	SFEED_PLUMB_INTERACTIVE=0 SFEED_PLUMB=thescript sfeed_curses ~/.sfeed/feeds/*

The program now waits on input while in non-interactive mode and only seems to
hang.

After:

The program starts but just has no input passed to it.